Introduces a ConsoleService, moves some Target specific methods to the Target class, move services to their own folder
