+++ letsencrypt-win-simple/letsencrypt-win-simple.csproj
@@ -155,12 +155,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Configuration\AcmeClientService.cs" />
+    <Compile Include="Services\AcmeClientService.cs" />
+    <Compile Include="Services\ConsoleService.cs" />
     <Compile Include="DefaultRunner.cs" />
-    <Compile Include="Certificates\CertificateService.cs" />
+    <Compile Include="Services\CertificateService.cs" />
     <Compile Include="Configuration\App.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
-    <Compile Include="Certificates\LetsEncryptService.cs" />
+    <Compile Include="Services\LetsEncryptService.cs" />
     <Compile Include="Extensions\ObjectExtensions.cs" />
     <Compile Include="Plugin\WebDavPlugin.cs" />
     <Compile Include="Plugin\IISPlugin.cs" />
@@ -218,6 +219,7 @@
     <Analyzer Include="..\packages\AWSSDK.Route53.3.3.5\analyzers\dotnet\cs\AWSSDK.Route53.CodeAnalysis.dll" />
     <Analyzer Include="..\packages\AWSSDK.S3.3.3.5.7\analyzers\dotnet\cs\AWSSDK.S3.CodeAnalysis.dll" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\Microsoft.Bcl.Build.1.0.21\build\Microsoft.Bcl.Build.targets" Condition="Exists('..\packages\Microsoft.Bcl.Build.1.0.21\build\Microsoft.Bcl.Build.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
